11/tapsum2json/requirements.txt
tap.py==3.1
six==1.16.0
regex==2022.3.2